fix table warnings
